+++ Brio/Game/Scene/SceneService.cs
@@ -2,6 +2,7 @@ using Brio.Capabilities.Actor;
 using Brio.Capabilities.Posing;
 using Brio.Capabilities.World;
 using Brio.Config;
+using Brio.Core;
 using Brio.Entities;
 using Brio.Entities.Actor;
 using Brio.Entities.Core;
@@ -10,6 +11,7 @@ using Brio.Files;
 using Brio.Game.Actor;
 using Brio.Game.Actor.Appearance;
 using Brio.Game.Actor.Extensions;
+using Brio.Game.Actor.Interop;
 using Brio.Game.Camera;
 using Brio.Game.Core;
 using Brio.Game.Posing;
@@ -93,7 +95,7 @@ public class SceneService(EntityManager _entityManager, VirtualCameraManager _vi
                 );
             }
             else
-            {
+            { 
                 var (actorId, actor) = actorCapability.CreateCharacter(actorFile.HasChild, false, forceSpawnActorWithoutCompanion: !actorFile.HasChild);
 
                 _framework.RunUntilSatisfied(
@@ -156,6 +158,20 @@ public class SceneService(EntityManager _entityManager, VirtualCameraManager _vi
         }, delayTicks: 2);
     }
 
+    //private static void ImportShadersFromFile(ref ModelShaderOverride modelShaderOverride, BrioHuman.ShaderParams shaderParams)
+    //{
+    //    modelShaderOverride.SkinColor = shaderParams.SkinColor;
+    //    modelShaderOverride.SkinGloss = shaderParams.SkinGloss;
+    //    modelShaderOverride.MuscleTone = shaderParams.MuscleTone;
+    //    modelShaderOverride.MouthColor = shaderParams.MouthColor;
+    //    modelShaderOverride.HairColor = shaderParams.HairColor;
+    //    modelShaderOverride.HairGloss = shaderParams.HairGloss;
+    //    modelShaderOverride.HairHighlight = shaderParams.HairHighlight;
+    //    modelShaderOverride.LeftEyeColor = shaderParams.LeftEyeColor;
+    //    modelShaderOverride.RightEyeColor = shaderParams.RightEyeColor;
+    //    modelShaderOverride.FeatureColor = shaderParams.FeatureColor;
+    //}
+
     private async Task ApplyDataToActor(EntityId actorId, ActorFile actorFile)
     {
         var attachedActor = _entityManager.GetEntity<ActorEntity>(actorId)!;
@@ -169,6 +185,7 @@ public class SceneService(EntityManager _entityManager, VirtualCameraManager _vi
 
         await _framework.RunOnTick(async () =>
         {
+            BrioUtilities.ImportShadersFromFile(ref appearanceCapability._modelShaderOverride, actorFile.AnamnesisCharaFile);
             await appearanceCapability.SetAppearance(actorFile.AnamnesisCharaFile, AppearanceImportOptions.All);
 
             await _framework.RunOnTick(async () =>
