+++ Brio/UI/Controls/Stateless/FileUIHelpers.cs
@@ -274,20 +274,6 @@ public class FileUIHelpers
                 }, ConfigurationService.Instance.Configuration.LastExportPath, true);
     }
 
-    public static void ImportShadersFromFile(ref ModelShaderOverride modelShaderOverride, BrioHuman.ShaderParams shaderParams)
-    {
-        modelShaderOverride.SkinColor = shaderParams.SkinColor;
-        modelShaderOverride.SkinGloss = shaderParams.SkinGloss;
-        modelShaderOverride.MuscleTone = shaderParams.MuscleTone;
-        modelShaderOverride.MouthColor = shaderParams.MouthColor;
-        modelShaderOverride.HairColor = shaderParams.HairColor;
-        modelShaderOverride.HairGloss = shaderParams.HairGloss;
-        modelShaderOverride.HairHighlight = shaderParams.HairHighlight;
-        modelShaderOverride.LeftEyeColor = shaderParams.LeftEyeColor;
-        modelShaderOverride.RightEyeColor = shaderParams.RightEyeColor;
-        modelShaderOverride.FeatureColor = shaderParams.FeatureColor;
-    }
-
     public static void ShowImportCharacterModal(ActorAppearanceCapability capability, AppearanceImportOptions options)
     {
         List<Type> types = [typeof(ActorAppearanceUnion), typeof(AnamnesisCharaFile)];
@@ -310,7 +296,7 @@ public class FileUIHelpers
                     if (options.HasFlag(AppearanceImportOptions.Shaders))
                     {
                         BrioHuman.ShaderParams shaderParams = appearanceFile;
-                        ImportShadersFromFile(ref capability._modelShaderOverride, shaderParams);
+                        BrioUtilities.ImportShadersFromFile(ref capability._modelShaderOverride, shaderParams);
                     }
                     _ = capability.SetAppearance(appearanceFile, options);
                 }
